// src/types/resources.ts
export type HeaderResourcesProps = {
  texasChicken: string;
  langBtn: string;
  change: string;
  findStore: string;
  signIn: string;
  goodMorning: string;
  goodAfternoon: string;
  goodEvening: string;
  myCart: string;
  profile: string;
  rewards: string;
  orderHistory: string;
  favourites: string;
  logOut: string;
  searchHere: string;
  trackOrder: string;
  hiGuest: string;
};

export type MenuSearchResourcesProps = {
  searchHere: string;
};

export type FooterResourcesProps = {
  footerDesc: string;
  texasChicken: string;
  change: string;
  discover: string;
  followUsOn: string;
  letsTalk: string;
  downloadOurApps: string;
  downloadOn: string;
  googlePlayStore: string;
  appStore: string;
  privacyPolicy: string;
  termsConditions: string;
  texasWay: string;
  menu: string;
  texas: string;
  viewAll: string;
};

export type BottomNavResourcesProps = {
  home: string;
  menu: string;
  login: string;
  profile: string;
  more: string;
  settings: string;
  langBtn: string;
  country: string;
  contactUs: string;
  termsConditions: string;
  privacyRequest: string;
  faq: string;
  reportIssue: string;
  ourStory: string;
  locations: string;
  halal: string;
  helpCenter: string;
  news: string;
  blogs: string;
  careers: string;
  party: string;
  franchising: string;
  trackOrder: string;
  suggestFeature: string;
  privacyPolicy: string;
  getApp: string;
  birthdayPackage: string;
  rewards: string;
};